requirements.txt
flask==2.3.3
flask-restx==1.3.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
jinja2==3.1.2
GitPython==3.1.40
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
pydantic==2.5.0
PyYAML==6.0.1
gunicorn==21.2.0